✏️  Decomposing S3 bench code
