+++ Cargo.toml
@@ -44,8 +44,14 @@ aws-sdk-s3 = { version = "1.78.0" }
 criterion = { version = "0.7.0", features = ["async_tokio"] }
 walkdir = { version = "2.5.0" }
 
+[[bench]]
+name = "shared"
+path = "benches/shared.rs"
+test = false # shared benchmark code, not meant to be ran alone
 
 [[bench]]
 name = "s3_bench"
+path = "benches/s3_bench.rs"
 harness = false
 required-features = ["icechunk"]
+
